pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "code-debug-env"
version = "1.0.0"
description = "OpenEnv: debug Python code until tests pass"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openenv-core>=0.2.1",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.0.0",
  "pytest>=8.0.0",
  "openai>=1.0.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[tool.setuptools]
py-modules = ["code_debug_env", "tasks", "reward", "inference"]
